requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.6.1               # via kombu
appdirs==1.4.4            # via black
asgiref==3.3.1            # via django
atomicwrites==1.4.0       # via pytest
attrs==20.3.0             # via pytest
backcall==0.2.0           # via ipython
beautifulsoup4==4.9.3     # via django-google-analytics-app
billiard==3.6.3.0         # via celery
black==20.8b1             # via -r requirements.in
celery==4.4.7             # via django-google-analytics-app
certifi==2020.12.5        # via requests
cffi==1.14.4              # via cryptography
chardet==4.0.0            # via requests
click==7.1.2              # via black, pip-tools
colorama==0.4.4           # via ipython, pytest
cryptography==3.3.1       # via pyjwt
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via python3-openid
django-allauth==0.44.0    # via -r requirements.in
django-cacheops==5.1      # via -r requirements.in
django-configurations==2.2  # via -r requirements.in
django-crispy-forms==1.10.0  # via -r requirements.in
django-debug-toolbar==3.2  # via -r requirements.in
django-extensions==3.1.0  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django-google-analytics-app==5.0.1  # via -r requirements.in
django-mysql==3.10.0      # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-allauth, django-cacheops, django-debug-toolbar, django-filter, django-google-analytics-app, django-mysql
envdir==1.0.1             # via -r requirements.in
flake8==3.8.4             # via -r requirements.in
funcy==1.15               # via django-cacheops
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
ipykernel==5.4.2          # via -r requirements.in
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in, ipykernel
jedi==0.17.2              # via ipython
jupyter-client==6.1.7     # via ipykernel
jupyter-core==4.7.0       # via jupyter-client
kiwisolver==1.3.1         # via matplotlib
kombu==4.6.11             # via celery
matplotlib==3.3.3         # via seaborn
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black
mysql-connector-python==8.0.22  # via -r requirements.in
mysqlclient==2.0.2        # via -r requirements.in
numpy==1.19.3             # via matplotlib, pandas, scipy, seaborn
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.8           # via pytest
pandas==1.1.5             # via -r requirements.in, seaborn
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via matplotlib
pip-tools==5.4.0          # via -r requirements.in
plotly==4.14.1            # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.8     # via ipython
protobuf==3.14.0          # via mysql-connector-python
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pygments==2.7.3           # via ipython
pyjwt[crypto]==2.0.0      # via django-allauth
pyparsing==2.4.7          # via matplotlib, packaging
pytest==6.2.1             # via -r requirements.in
python-dateutil==2.8.1    # via jupyter-client, matplotlib, pandas
python3-openid==3.2.0     # via django-allauth
pytz==2020.4              # via celery, django, pandas
pywin32==300              # via jupyter-core
pyzmq==20.0.0             # via jupyter-client
redis==3.5.3              # via django-cacheops
regex==2020.11.13         # via black
requests-oauthlib==1.3.0  # via django-allauth
requests==2.25.1          # via django-allauth, django-google-analytics-app, requests-oauthlib
retrying==1.3.3           # via plotly
scipy==1.5.4              # via seaborn
seaborn==0.11.1           # via -r requirements.in
six==1.15.0               # via cryptography, cycler, django-cacheops, django-configurations, django-google-analytics-app, pip-tools, plotly, protobuf, python-dateutil, retrying, structlog
soupsieve==2.1            # via beautifulsoup4
sqlalchemy==1.3.22        # via -r requirements.in
sqlparse==0.4.1           # via django, django-debug-toolbar
structlog==20.1.0         # via django-google-analytics-app
toml==0.10.2              # via black, pytest
tornado==6.1              # via ipykernel, jupyter-client
traitlets==5.0.5          # via ipykernel, ipython, jupyter-client, jupyter-core
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.26.2           # via requests
vine==1.3.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit
whitenoise==5.2.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
